% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/increment_counter.R
\name{dn_figure_rename}
\alias{dn_figure_rename}
\title{figure rename}
\usage{
dn_figure_rename(counter, identifier = NULL)
}
\arguments{
\item{counter}{the counter object}

\item{identifier}{the identifier for the figure}
}
\value{
character
}
\description{
Generate a figure number for renaming within a path.
}
