% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/increment_counter.R
\name{dn_figure_string}
\alias{dn_figure_string}
\title{figure string}
\usage{
dn_figure_string(counter, identifier = NULL, supplemental = FALSE)
}
\arguments{
\item{counter}{the counter object}

\item{identifier}{the identifier to get the number for}

\item{supplemental}{should an "S" be added to the label first?}
}
\value{
character
}
\description{
Generate a figure string with the number included.
}
