% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/increment_counter.R
\name{dn_increment_counter}
\alias{dn_increment_counter}
\title{increment counter}
\usage{
dn_increment_counter(counter, name)
}
\arguments{
\item{counter}{the counter to add to}

\item{name}{the name for the value}
}
\value{
named numeric
}
\description{
given a counter, add to it with a specific name
}
