% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_counters.R
\name{dn_initialize_counter}
\alias{dn_initialize_counter}
\title{initialize counters}
\usage{
dn_initialize_counter()
}
\value{
character vector
}
\description{
Sets up a vector to keep track of counting things you want.
}
