% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/increment_counter.R
\name{dn_modify_path}
\alias{dn_modify_path}
\title{modify figure path}
\usage{
dn_modify_path(figure_path, options)
}
\arguments{
\item{figure_path}{the original figure path}

\item{options}{the chunk options}
}
\value{
character
}
\description{
This function uses chunk options to rename the figure file.
You should add a chunk option called \code{counter_identifier} that uses
\code{dn_figure_rename} to get the figure number. See the vignette for an
example of how to use this in an actual document.
}
\details{
We often want to include the figure number in figure file names.
This function allows us to rename them easily.
}
