% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/increment_counter.R
\name{dn_paste_counter}
\alias{dn_paste_counter}
\title{paste counter}
\usage{
dn_paste_counter(counter, pre_string = "", identifier = NULL)
}
\arguments{
\item{counter}{the counter to search on}

\item{pre_string}{any text to paste first}

\item{identifier}{the id for the count}
}
\value{
string
}
\description{
Given a counter and a name to find, returns a string for insertion
into the text.
}
