% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/increment_counter.R
\name{dn_table_string}
\alias{dn_table_string}
\title{table string}
\usage{
dn_table_string(counter, identifier = NULL)
}
\arguments{
\item{counter}{the counter object}

\item{identifier}{the identifier to get the number for}
}
\value{
character
}
\description{
Generate a "Table " string with the number included.
}
